+++ resources/views/Consignee/Joborders/joborderrequests.blade.php
@@ -108,5 +108,11 @@
             @endforeach
         </div>
     @else
+    
+        <div class="col-12">
+            <div class="alert alert-danger text-center">
+                <i class="fas fa-exclamation-triangle mr-2"></i>NO PENDING JOB ORDERS!
+            </div>
+        </div>
     @endif
 </div>