+++ resources/views/ForwardRequests/scripts.blade.php
@@ -1,0 +1,1 @@
+    <script src="/others/stisla_admin_v1.0.0/dist/js/forwardrequests/forwardrequests.js"></script>